requirements.txt
Flask==2.1.0
requests==2.26.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.1.0
Werkzeug==2.1.0
python-dotenv==0.18.0
pandas==1.3.3
stopwordsiso==0.6.1
scikit-learn
nltk==3.6.2
